+++ app/src/domain/fetch/factories/ServiceFactory.ts
@@ -1,9 +1,9 @@
-import { DIContainer, SERVICE_TOKENS } from "../core/di/Container";
-import { ILogger, LoggerFactory } from "../core/logger/Logger";
-import { ICache, CacheFactory } from "../core/cache/Cache";
-import { RetryManager, RetryManagerFactory } from "../core/retry/RetryManager";
-import { ConfigManager } from "../core/config/Config";
-import { ValidatorFactory } from "../core/validation/Validator";
+import { DIContainer, SERVICE_TOKENS } from "../../shared/di/Container";
+import { ILogger, LoggerFactory } from "../../shared/logger/Logger";
+import { ICache, CacheFactory } from "../../shared/cache/Cache";
+import { RetryManager, RetryManagerFactory } from "../../shared/retry/RetryManager";
+import { ConfigManager } from "../../shared/config/Config";
+import { ValidatorFactory } from "../../shared/validation/Validator";
 import { NotionDatabaseService } from "../services/NotionDatabaseService";
 import { HabitMapper } from "../mappers/HabitMapper";
 import { HabitRepository } from "../repositories/HabitRepository";
