+++ app/src/domain/fetch/services/NotionDatabaseService.ts
@@ -1,10 +1,10 @@
 import { notionClient } from "../../../lib/notionhq/init";
 import { DatabaseResponse, BlockObjectResponse } from "../../../lib/notionhq/type";
-import { FetchError, ERROR_CODES } from "../core/errors/FetchError";
-import { ILogger } from "../core/logger/Logger";
-import { ICache } from "../core/cache/Cache";
-import { RetryManager } from "../core/retry/RetryManager";
-import { ValidatorFactory } from "../core/validation/Validator";
+import { FetchError, ERROR_CODES } from "../../shared/errors/FetchError";
+import { ILogger } from "../../shared/logger/Logger";
+import { ICache } from "../../shared/cache/Cache";
+import { RetryManager } from "../../shared/retry/RetryManager";
+import { ValidatorFactory } from "../../shared/validation/Validator";
 
 /**
  * Notionデータベース操作のためのサービスクラス
