+++ app/src/domain/shared/errors/FetchError.ts
